LayoutTests/fast/canvas/offscreen-disabled-construct3-quirk.html

<!-- webkit-test-runner [ NeedsSiteSpecificQuirks=true ] -->
<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Test to ensure that OffscreenCanvas is disabled when the Construct3 library is present.");

window.C3_IsSupported = true;
if (typeof OffscreenCanvas !== "undefined") {
    testFailed("OffscreenCanvas is valid and the feature is not disabled");
} else {
    testPassed("OffscreenCanvas is undefined and the feature is disabled");
}

</script>
</body>
</html>
