LayoutTests/fast/canvas/offscreen-disabled-construct3-quirk-expected.txt

Test to ensure that OffscreenCanvas is disabled when the Construct3 library is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS OffscreenCanvas is undefined and the feature is disabled
PASS successfullyParsed is true

TEST COMPLETE
